# lib/api-zod/src/generated/types/columnInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ColumnInfo {
  name: string;
  type: string;
  description: string;
  primaryKey: boolean;
  nullable: boolean;
}