# lib/api-zod/src/generated/types/generateSQLRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateSQLRequest {
  /** Natural language question to convert to SQL */
  query: string;
  /**
   * Maximum number of generate→validate loops (default 5, max 10)
   * @minimum 1
   * @maximum 10
   */
  maxIterations?: number;
}