# lib/api-zod/src/generated/types/generateSQLResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateSQLResponse {
  /** Original user query */
  query: string;
  /** Generated SAP HANA SQL query */
  sql: string;
  /** Whether the validator approved the final SQL */
  passed: boolean;
  /** Number of generate→validate iterations performed */
  iterations: number;
  /** Validator feedback if the final query did not pass */
  feedback?: string | null;
  /** Detailed step-by-step log of all agent actions */
  agentLog: string[];
  /** Error message if the pipeline encountered a fatal error */
  error?: string | null;
}