# lib/api-zod/src/generated/types/listTablesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TableInfo } from "./tableInfo";

export interface ListTablesResponse {
  tables: TableInfo[];
}