# lib/api-zod/src/generated/types/tableInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ColumnInfo } from "./columnInfo";

export interface TableInfo {
  schema: string;
  tableName: string;
  fullName: string;
  description: string;
  columns: ColumnInfo[];
}